Fixing KonaneMove to delete unecessary functions
