--- Sources/File.swift
@@ -1,9 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Alex Hill on 10/26/16.
-//
-//
-
-import Foundation
