+++ Sources/KonaneGame.swift
@@ -52,94 +52,8 @@ class KonaneGame {
 
 
 func play() {
-    
-    //First build the game board
-    populateGameBoard()
-    
-    //function to print the game board
-    displayBoard()
-    
-    //function to remove pieces - based on input source?
-    
-    
-    //function to jump a piecce
-        //if no piece is available to jump, return bool: whiteHasWon: true or blackHasWon: true
+    //ADD CONTENTS AND ACTUALLY WRITE CODE
 }
 
 
-
-//BEGIN Populate gameBoard
-func populateGameBoard() {
-    for column in 0..<gameState.width {
-        gameState.gameBoard.append([]) //Each column is an array
-        for row in 0..<gameState.height {
-            if column % 2 == 0 { //If column even -> black first
-                if row % 2 == 0 {
-                    gameState.gameBoard[column].append(KonaneColor.black)
-                }
-                else {
-                    gameState.gameBoard[column].append(KonaneColor.white)
-                }
-            }
-            else {
-                if row % 2 == 0 { //If column odd -> white first
-                    gameState.gameBoard[column].append(KonaneColor.white)
-                }
-                else {
-                    gameState.gameBoard[column].append(KonaneColor.black)
-                }
-            }
-        }
-    }
-}
-//END Populate gameBoard
-
-
-
-
-//Displays the game board in terminal
-//FIX WITH PROPER VARIABLE NAMES
-func displayBoard() {
-    
-    var arbitraryVariable = 0
-    
-    for rowNumber in (0..<gameState.height).reversed() {
-        //Print row numbering
-        if rowNumber < 10 {
-            print(" \(rowNumber)", terminator:" ")
-        }
-        else {
-            print("\(rowNumber)", terminator:" ")
-        }
-        //Print actual board
-        for columnNumber in 0..<gameState.width {
-            if gameState.gameBoard[columnNumber][rowNumber] == KonaneColor.black {
-                print(" x", terminator: " ") //black square
-            }
-            else if gameState.gameBoard[columnNumber][rowNumber] == KonaneColor.white {
-                print(" o", terminator: " ") //white square
-            }
-            else {
-                print("  ", terminator: " ") //empty square
-            }
-        }
-        print() //uses a terminator to make new line
-        //Print column numbering
-        
-    }
-    
-    print("   ", terminator:"")
-    while arbitraryVariable < gameState.height {
-        if arbitraryVariable < 10 {
-            print(" \(arbitraryVariable)", terminator: " ")
-        } else {
-            print("\(arbitraryVariable)", terminator: " ")
-            
-        }
-        arbitraryVariable += 1
-    }
-    
-    print()
-}
-//END printGameBoard()
